+++ CHANGES.rst
@@ -2,6 +2,11 @@
  Breaking changes
 ==================
 
+Current
+=======
+
+- Change the cache format to support multi-language caches.
+
 Version 1.1
 ===========